messaging_app/settings.py
import os
import environ

# Initialize env reader
env = environ.Env()
environ.Env.read_env()

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.mysql',
        'NAME': env('MYSQL_DB'),
        'USER': env('MYSQL_USER'),
        'PASSWORD': env('MYSQL_PASSWORD'),
        'HOST': 'db',  # this is the service name used in docker-compose
        'PORT': '3306',
    }
}
