messaging_app/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==25.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
django-environ==0.12.0
mysqlclient==2.2.4
